Add variable for additional exposed Ports/Services
Since Fluent-bit supports network inputs (TCP and in the future UDP) it makes sense that the helm chart have the ability to expose additional ports at both the daemonset and the service level.

This PR adds an additional variable extraPorts which is a list of objects, and the different helm templates will use the variable values to add additional exposed ports to the daemonset and service.

An example usage:
To expose TCP port 1234

The variable extraPorts can be set to:

extraPorts:
  - name: tcp
    protocol: TCP
    port: 1234
    containerPort: 1234
That will result in a service that looks like this:

.
.
.
  type: ClusterIP
  ports:
    - port: 2020
      targetPort: http
      protocol: TCP
      name: http
    - name: tcp
      targetPort: http2
      protocol: TCP
      port: 1234
  selector:
    app.kubernetes.io/name: fluent-bit
.
.
.
and a daemonset which looks like:

.
.
.
          image: "fluent/fluent-bit:1.4.4"
          imagePullPolicy: Always
          ports:
            - name: http
              containerPort: 2020
              protocol: TCP
            - name: tcp
              containerPort: 1234
              protocol: TCP
          livenessProbe:
            httpGet:
              path: /
              port: http
.
.
.
The above are a subset of the actual files, but they show how that one variable extraPorts can be used to expose additional ports.

Signed-off-by: Justin Witrick <jwitrick@gmail.com>