feat(fluent-bit): Updated image to v3.1.10
The following PR updates the Fluent Bit image to the latest version
available: v3.1.10.

In addition, I am reverting the change that was done in the PullPolicy
for the main image from 'IfNotPresent' to 'Always'. As you know, we
don't do user-telemetry in Fluent Bit, our only indicator of adoption
is the number of deployments/downloads from the containers registry.
Users can always tweak this value, so I am just changing the default.

Signed-off-by: Eduardo Silva <eduardo@calyptia.com>